enumerator: windows: Rewrite USB device hotpluging to support composite devices
In Windows 7/8, our TinyG board shows up as a composite device with two
endpoints. Due to unfathomable reasons, Windows does not send out device
notifications when the endpoints appear, and instead only sends a single
notification that the parent composite device has appeared.

This patch replaces all handling of incremental device add/remove notifications
with a brute force approach. Every time the device tree changes, we wait 100ms
for devices to settle then rescan all devices to find which ones are supported
and emit signals to reflect the change in state.

This is fragile and intended as a stopgap solution until more robust cross
platform hardware detection is implemented with libusb.